Change error code type to uint8

+++ modbus/modbus-mduino/OctaveModbusWrapper.h
@@ -14,7 +14,7 @@
 #include "ParamTables.h"
 
 // Bit indices to check for alarms
-const int alarmsIndices[] = {0, 5, 7, 11, 12, 13};
+const uint8_t alarmsIndices[] = {0, 5, 7, 11, 12, 13};
 
 // Scale factor for two decimal places
 #define SCALE_FACTOR "100.0"
@@ -32,67 +32,67 @@ class OctaveModbusWrapper {
         void begin();
         void InitMaps();
 
-        int AwaitResponse();
+        uint8_t AwaitResponse();
         void ProcessResponse(ModbusResponse *response);
-        int BlockingReadRegisters(int startMemAddress, int numValues, int signedValueSizeinBits);
-        int BlockingWriteSingleRegister(int memAddress, int value);
+        uint8_t BlockingReadRegisters(uint8_t startMemAddress, uint8_t numValues, int8_t signedValueSizeinBits);
+        uint8_t BlockingWriteSingleRegister(uint8_t memAddress, int16_t value);
 
-        int ReadAlarms();
-        int SerialNumber();
-        int ReadWeekday();
-        int ReadDay();
-        int ReadMonth();
-        int ReadYear();
-        int ReadHours();
-        int ReadMinutes();
-        int VolumeUnit();
-        int ForwardVolume(int unsignedValueSizeinBits);
-        int ReverseVolume(int unsignedValueSizeinBits);
-        int ReadVolumeResIndex();
-        int SignedCurrentFlow(int unsignedValueSizeinBits);
-        int ReadFlowResIndex();
-        int FlowUnit();
-        int FlowDirection();
-        int TemperatureValue();
-        int TemperatureUnit();
-        int NetSignedVolume(int unsignedValueSizeinBits);
-        int NetUnsignedVolume(int unsignedValueSizeinBits);
-        int SystemReset();
-        int WriteWeekday(int value);
-        int WriteDay(int value);
-        int WriteMonth(int value);
-        int WriteYear(int value);
-        int WriteHours(int value);
-        int WriteMinutes(int value);
-        int WriteVolumeResIndex(int value);
-        int WriteFlowResIndex(int value);
+        uint8_t ReadAlarms();
+        uint8_t SerialNumber();
+        uint8_t ReadWeekday();
+        uint8_t ReadDay();
+        uint8_t ReadMonth();
+        uint8_t ReadYear();
+        uint8_t ReadHours();
+        uint8_t ReadMinutes();
+        uint8_t VolumeUnit();
+        uint8_t ForwardVolume(uint8_t unsignedValueSizeinBits);
+        uint8_t ReverseVolume(uint8_t unsignedValueSizeinBits);
+        uint8_t ReadVolumeResIndex();
+        uint8_t SignedCurrentFlow(uint8_t unsignedValueSizeinBits);
+        uint8_t ReadFlowResIndex();
+        uint8_t FlowUnit();
+        uint8_t FlowDirection();
+        uint8_t TemperatureValue();
+        uint8_t TemperatureUnit();
+        uint8_t NetSignedVolume(uint8_t unsignedValueSizeinBits);
+        uint8_t NetUnsignedVolume(uint8_t unsignedValueSizeinBits);
+        uint8_t SystemReset();
+        uint8_t WriteWeekday(uint8_t value);
+        uint8_t WriteDay(uint8_t value);
+        uint8_t WriteMonth(uint8_t value);
+        uint8_t WriteYear(uint8_t value);
+        uint8_t WriteHours(uint8_t value);
+        uint8_t WriteMinutes(uint8_t value);
+        uint8_t WriteVolumeResIndex(uint8_t value);
+        uint8_t WriteFlowResIndex(uint8_t value);
 
         /****** Modbus response buffers ******/
-        int int16Buffer[16];
+        int16_t int16Buffer[16];
         int32_t int32Buffer;
         uint32_t uint32Buffer;
         float64_t doubleBuffer;
 
         /****** Parameter maps ********/
-        std::map<String, int> flowUnitNameToCode;
-        std::map<int, String> flowUnitCodeToName;
-        std::map<String, int> volumeUnitNameToCode;
-        std::map<int, String> volumeUnitCodeToName;
-        std::map<String, int> temperatureUnitNameToCode;
-        std::map<int, String> temperatureUnitCodeToName;
-        std::map<String, int> flowDirectionNameToCode;
-        std::map<int, String> flowDirectionCodeToName;
-        std::map<String, int> resolutiontNameToCode;
-        std::map<int, String> resolutionCodeToName;
-        std::map<int, String> alarmCodeToName;
+        std::map<String, uint8_t> flowUnitNameToCode;
+        std::map<uint8_t, String> flowUnitCodeToName;
+        std::map<String, uint8_t> volumeUnitNameToCode;
+        std::map<uint8_t, String> volumeUnitCodeToName;
+        std::map<String, uint8_t> temperatureUnitNameToCode;
+        std::map<uint8_t, String> temperatureUnitCodeToName;
+        std::map<String, uint8_t> flowDirectionNameToCode;
+        std::map<uint8_t, String> flowDirectionCodeToName;
+        std::map<String, uint8_t> resolutiontNameToCode;
+        std::map<uint8_t, String> resolutionCodeToName;
+        std::map<uint8_t, String> alarmCodeToName;
 
         /****** Function name map *******/
-        std::map<String, int> functionNameToCode;
-        std::map<int, String> functionCodeToName;
+        std::map<String, uint16_t> functionNameToCode;
+        std::map<uint16_t, String> functionCodeToName;
 
-        std::map<int, String> errorCodeToName;
+        std::map<uint8_t, String> errorCodeToName;
 
-        int lastUsedFunctionCode = 0;
+        uint16_t lastUsedFunctionCode = 0;
 
         uint32_t modbusBaudrate = MODBUS_BAUDRATE;
     
@@ -102,9 +102,9 @@ class OctaveModbusWrapper {
 
         /****** Parameters for the Modbus requests ******/
         // Number of registers to read for a Modbus request, is 0 for a write request
-        int _numRegisterstoRead = 0;
+        uint8_t _numRegisterstoRead = 0;
         // Size, in bits, of the slave response values, is -32 for int32 and 32 for uint32
-        int _signedResponseSizeinBits = 16;
+        int8_t _signedResponseSizeinBits = 16;
 };
 
 #endif
